// types/type.d.ts
export type StringMap = Record<string, unknown>;

export type ActionItem = {
  name: string;
  onTap: StringMap;
  extern: StringMap;
};

export type ImageItem = {
  id: string;
  url: string;
  name: string;
  path: string;
  extern: StringMap;
};

export type MetadataListItem = {
  type: string;
  name: string;
  value: ActionItem[];
};

export type PagingInfo = {
  page: number;
  pages: number;
  total: number;
  hasReachedMax: boolean;
};

export type ComicListItem = {
  source: string;
  id: string;
  title: string;
  subtitle: string;
  finished: boolean;
  likesCount: number;
  viewsCount: number;
  updatedAt: string;
  cover: ImageItem;
  metadata: MetadataListItem[];
  raw: StringMap;
  extern: StringMap;
};